% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorability.R
\name{factorability}
\alias{factorability}
\title{factorability: Quick report for KMO and Bartlett's test results}
\usage{
factorability(data, digits = 2)
}
\arguments{
\item{data}{A data frame or tibble with numeric variables to assess factorability.}

\item{digits}{Number of decimal places to round KMO and chi-squared values. Default is 2.}
}
\value{
No return value. Prints a concise report to the console with:
\itemize{
\item Overall KMO value
\item Bartlett’s test chi-squared value, degrees of freedom, and formatted p-value
}
}
\description{
This function computes and prints a quick summary of the Kaiser-Meyer-Olkin (KMO)
Measure of Sampling Adequacy and Bartlett’s Test of Sphericity to assess whether
data are appropriate for exploratory factor analysis (EFA).
}
\examples{
\dontrun{
factorability(mtcars[, 1:5])
}

}
