% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prty_confmat.R
\name{prty_confmat}
\alias{prty_confmat}
\title{Pretty Confusion Matrix (Predicted in Columns)}
\usage{
prty_confmat(predicted, actual, group_name = "Model")
}
\arguments{
\item{predicted}{The predicted labels.}

\item{actual}{The actual observed labels.}
}
\value{
A confusion matrix table.
}
\description{
Create a nicely formatted confusion matrix where Predicted classes are in columns and Observed classes are in rows.
}
