% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{ds: Descriptive statistics table}
\usage{
ds(data, decimals = 2, labels = NULL)
}
\arguments{
\item{data}{A data frame containing numeric variables.}

\item{decimals}{Optional decimals to return. If \code{NULL} (default), 2 decimals are used.}

\item{labels}{Optional labels for each variable.}
}
\value{
A tibble with columns: \code{var}, \code{N}, \verb{Mean ± SD}, \verb{Median ± MAD}, and \code{IQR}.
}
\description{
Creates a table with neat descriptive statistics.
}
